RenderContext.

+++ Source/Platform/Window.cpp
@@ -7,7 +7,7 @@
 
 #include "Window.hpp"
 
-Window::Window(std::string name, int width, int height)
+Window::Window(std::string name, size_t width, size_t height)
 :name(name), width(width), height(height){
     glfwInit();
     glfwWindowHint(GLFW_CLIENT_API, GLFW_NO_API);
